\encoding{utf8}
\name{heinrichi}
\alias{heinrichi}
\docType{data}
\title{Heinrichi data set}
\description{Data from the artificial textual tradition \emph{Heinrichi}}
\usage{data(heinrichi)}
\format{
  A matrix with 1208 observations on the following 37 variables.
  \describe{
    \item{\code{A}}{a numeric vector}
    \item{\code{Ab}}{a numeric vector}
    \item{\code{Ac}}{a numeric vector}
    \item{\code{Ad}}{a numeric vector}
    \item{\code{Ae}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{Ba}}{a numeric vector}
    \item{\code{Bb}}{a numeric vector}
    \item{\code{Bd}}{a numeric vector}
    \item{\code{Be}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{Ca}}{a numeric vector}
    \item{\code{Cb}}{a numeric vector}
    \item{\code{Cc}}{a numeric vector}
    \item{\code{Cd}}{a numeric vector}
    \item{\code{Ce}}{a numeric vector}
    \item{\code{Cf}}{a numeric vector}
    \item{\code{Da}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
    \item{\code{F}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{H}}{a numeric vector}
    \item{\code{I}}{a numeric vector}
    \item{\code{J}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{L}}{a numeric vector}
    \item{\code{M}}{a numeric vector}
    \item{\code{N}}{a numeric vector}
    \item{\code{O}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{R}}{a numeric vector}
    \item{\code{S}}{a numeric vector}
    \item{\code{T}}{a numeric vector}
    \item{\code{V}}{a numeric vector}
    \item{\code{W}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{Z}}{a numeric vector}
  }
}
\details{
The data comes from an artificial tradition, created under controlled circumstances.
The data is presented here as used in Camps & Cafiero 2015, without further modifications or corrections.
Readings have been converted to numeric codes (0 being omission, and NA an absence of value).
}
\source{
Roos, Teemu, and Tuomas Heikkilä. ‘Evaluating methods for computer-assisted stemmatology using artificial benchmark data sets’. \emph{Literary and Linguistic Computing} 24/4 (2009), p. 417–433.

Roos, Teemu, Tuomas Heikkilä, and Petri Myllymäki. ‘Computer-Assisted Stemmatology Challenge’. Helsinki, 2007, \url{https://www.cs.helsinki.fi/u/ttonteri/casc/data.html}.
}
\references{
Camps, Jean-Baptiste, and Florian Cafiero. ‘Genealogical Variant Locations and Simplified Stemma: A Test Case’. \emph{Analysis of Ancient and Medieval Texts and Manuscripts: Digital Approaches}, edited by Tara Andrews and Caroline Macé, Brepols, 2015, pp. 69–93, \url{https://halshs.archives-ouvertes.fr/halshs-01435633}, DOI: \href{http://dx.doi.org/10.1484/M.LECTIO-EB.5.102565}{10.1484/M.LECTIO-EB.5.102565}.
}
\examples{
data(heinrichi)
}
\keyword{datasets}
