\encoding{utf8}
\name{notreBesoin}
\alias{notreBesoin}
\docType{data}
\title{Notre Besoin data set}
\description{Data from the artificial tradition \emph{Notre-Besoin}}
\usage{data(notreBesoin)}
\format{
  A matrix with 42 observations on the following 13 variables.
  \describe{
    \item{\code{nb1}}{a numeric vector}
    \item{\code{nb2}}{a numeric vector}
    \item{\code{nb3}}{a numeric vector}
    \item{\code{nb4}}{a numeric vector}
    \item{\code{nb5}}{a numeric vector}
    \item{\code{nb6}}{a numeric vector}
    \item{\code{nb7}}{a numeric vector}
    \item{\code{nb8}}{a numeric vector}
    \item{\code{nb9}}{a numeric vector}
    \item{\code{nb10}}{a numeric vector}
    \item{\code{nb11}}{a numeric vector}
    \item{\code{nb12}}{a numeric vector}
    \item{\code{nb13}}{a numeric vector}
  }
}
\details{%
The data comes from an artificial tradition, created under controlled circumstances.
The variant locations have been selected to retain only substantial readings.
The data is presented here as used in Camps & Cafiero 2015, without further modifications or corrections.
Readings have been converted to numeric codes (0 being omission, and NA an absence of value).
}
\source{
Baret, Philippe V., Robinson, P., and Macé, C., « Testing methods on an artificially created textual tradition », Linguistica computazionale 24 (2004), P. 1000–1029.

Roos, Teemu, Heikkilä, Tuomas, and Myllymäki, Petri, « Computer-Assisted Stemmatology Challenge », 2007, \url{https://www.cs.helsinki.fi/u/ttonteri/casc/data.html}.
}
\references{
Camps (Jean-Baptiste) & Cafiero (Florian), « Genealogical variant locations and simplified stemma: a test case », in Analysis of Ancient and Medieval Texts and Manuscripts: Digital Approaches, ed. Tara Andrews & Caroline Macé, Turnhout: Brepols, 2015, p. 69‑93 (Lectio, 1), \url{http://dx.doi.org/10.1484/M.LECTIO-EB.5.102565}.
}
\examples{
data(notreBesoin)
}
\keyword{datasets}
