\encoding{utf8}
\name{otinel}
\alias{otinel}
\docType{data}
\title{Otinel data set}
\description{Data for the tradition of the \emph{Chanson d'Otinel}}
\usage{data(otinel)}
\format{
  A matrix with 36 observations on the following 8 variables.
  \describe{
    \item{\code{A}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{M}}{a numeric vector}
    \item{\code{NOt}}{a numeric vector}
    \item{\code{WOt}}{a numeric vector}
    \item{\code{EOtA}}{a numeric vector}
    \item{\code{EOtF}}{a numeric vector}
    \item{\code{EOtT}}{a numeric vector}
  }
}
\details{
A sample of 36 substantial readings, taken from the tradition (Old French and translations) of the \emph{Chanson d'Otinel}. Data will be subject to updates.
Readings have been converted to numeric codes (0 being omission, and NA an absence of value).
}
\source{%
Camps, Jean-Baptiste. \emph{La `Chanson d’Otinel’: édition complète du corpus manuscrit et prolégomènes à l’édition critique}, PhD thesis, dir. Dominique Boutet, Paris-Sorbonne, 2016, \url{http://www.theses.fr/2016PA040173}.
}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(otinel)
}
\keyword{datasets}
