\encoding{utf8}
\name{parzival}
\alias{parzival}
\docType{data}
\title{Parzival data set}
\description{Data from the artificial textual tradition \emph{Parzival}}
\usage{data(parzival)}
\format{
  A matrix with 139 observations on the following 16 variables.
  \describe{
    \item{\code{p1}}{a numeric vector}
    \item{\code{p2}}{a numeric vector}
    \item{\code{p3}}{a numeric vector}
    \item{\code{p4}}{a numeric vector}
    \item{\code{p5}}{a numeric vector}
    \item{\code{p6}}{a numeric vector}
    \item{\code{p7}}{a numeric vector}
    \item{\code{p8}}{a numeric vector}
    \item{\code{p9}}{a numeric vector}
    \item{\code{p10}}{a numeric vector}
    \item{\code{p11}}{a numeric vector}
    \item{\code{p12}}{a numeric vector}
    \item{\code{p13}}{a numeric vector}
    \item{\code{p14}}{a numeric vector}
    \item{\code{p15}}{a numeric vector}
    \item{\code{p16}}{a numeric vector}
  }
}
\details{
The data comes from an artificial tradition, created under controlled circumstances.
The variant locations have been selected to retain only substantial readings.
The data is presented here as used in Camps & Cafiero 2015, without further modifications or corrections.
Readings have been converted to numeric codes (0 being omission, and NA an absence of value).
}
\source{
M. Spencer, E. A. Davidson, A. C. Barbrook, and C. J. Howe. ‘Phylogenetics of artificial manuscripts’. \emph{Journal of Theoretical Biology}, 227 (2004), p. 503–11, \url{https://doi.org/10.1016/j.jtbi.2003.11.022}.

Roos, Teemu, Tuomas Heikkilä, and Petri Myllymäki. ‘Computer-Assisted Stemmatology Challenge’. Helsinki, 2007, \url{https://www.cs.helsinki.fi/u/ttonteri/casc/data.html}.
}
\references{
Camps, Jean-Baptiste, and Florian Cafiero. ‘Genealogical Variant Locations and Simplified Stemma: A Test Case’. \emph{Analysis of Ancient and Medieval Texts and Manuscripts: Digital Approaches}, edited by Tara Andrews and Caroline Macé, Brepols, 2015, pp. 69–93, \url{https://halshs.archives-ouvertes.fr/halshs-01435633}, DOI: \href{http://dx.doi.org/10.1484/M.LECTIO-EB.5.102565}{10.1484/M.LECTIO-EB.5.102565}.
}
\examples{
data(parzival)
}
\keyword{datasets}
