\encoding{utf8}
\name{notreBesoin}
\alias{notreBesoin}
\docType{data}
\title{Notre Besoin data set}
\description{Data from the artificial tradition \emph{Notre-Besoin}}
\usage{data(notreBesoin)}
\format{
  A matrix with 42 observations on the following 13 variables.
  \describe{
    \item{\code{nb1}}{a numeric vector}
    \item{\code{nb2}}{a numeric vector}
    \item{\code{nb3}}{a numeric vector}
    \item{\code{nb4}}{a numeric vector}
    \item{\code{nb5}}{a numeric vector}
    \item{\code{nb6}}{a numeric vector}
    \item{\code{nb7}}{a numeric vector}
    \item{\code{nb8}}{a numeric vector}
    \item{\code{nb9}}{a numeric vector}
    \item{\code{nb10}}{a numeric vector}
    \item{\code{nb11}}{a numeric vector}
    \item{\code{nb12}}{a numeric vector}
    \item{\code{nb13}}{a numeric vector}
  }
}
\details{%
The data comes from an artificial tradition, created under controlled circumstances.
The variant locations have been selected to retain only substantial readings.
The data is presented here as used in Camps & Cafiero 2015, without further modifications or corrections.
Readings have been converted to numeric codes (0 being omission, and NA an absence of value).
}
\source{
Baret, Philippe V.,  P. Robinson, and C. Macé. ‘Testing methods on an artificially created textual tradition’. \emph{Linguistica computazionale} 24 (2004), p. 1000–1029.

Roos, Teemu, Tuomas Heikkilä, and Petri Myllymäki. ‘Computer-Assisted Stemmatology Challenge’. Helsinki, 2007, \url{https://www.cs.helsinki.fi/u/ttonteri/casc/data.html}.
}
\references{
Camps, Jean-Baptiste, and Florian Cafiero. ‘Genealogical Variant Locations and Simplified Stemma: A Test Case’. \emph{Analysis of Ancient and Medieval Texts and Manuscripts: Digital Approaches}, edited by Tara Andrews and Caroline Macé, Brepols, 2015, pp. 69–93, \url{https://halshs.archives-ouvertes.fr/halshs-01435633}, DOI: \href{http://dx.doi.org/10.1484/M.LECTIO-EB.5.102565}{10.1484/M.LECTIO-EB.5.102565}.
}
\examples{
data(notreBesoin)
}
\keyword{datasets}
